// src/services/models/torrent.ts
export type Torrent = {
  Name: string;
  DetailsUrl: string;
  OriginalName: string;
  RussianName: string;
  Year: string;
  Size: number;
  Magnet: string;
  Date: string;
  K4: boolean;
  FHD: boolean;
  HDR: boolean;
  HDR10: boolean;
  HDR10plus: boolean;
  DV: boolean;
  Seeds: number;
  Leeches: number;
  Hash: string;
  AvailabilityScore?: number;
  Categories?: string[];
  CategoryLabels?: string[];
  Peers?: number;
  Quality?: number;
  QualityLabel?: string;
  Seasons?: number[];
  SizeName?: string;
  Tracker?: string;
  UpdatedDate?: string;
  VideoType?: string;
  Voices?: string[];
};